update(internal): move notification logic to domain
